requirements.txt
blinker==1.4
docutils==0.14
feedgenerator==1.9
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysql-connector-python==8.0.13
pelican==3.7.1
prettytable==0.7.2
protobuf==3.6.1
psycopg2==2.7.6.1
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
Unidecode==1.0.22
